Refactors the addSignature test methods
