+++ rskj-core/src/test/java/co/rsk/peg/BridgeSupportTest.java
@@ -4,7 +4,6 @@ import co.rsk.bitcoinj.core.*;
 import co.rsk.bitcoinj.crypto.TransactionSignature;
 import co.rsk.bitcoinj.script.Script;
 import co.rsk.bitcoinj.script.ScriptBuilder;
-import co.rsk.bitcoinj.script.ScriptChunk;
 import co.rsk.bitcoinj.store.BlockStoreException;
 import co.rsk.blockchain.utils.BlockGenerator;
 import co.rsk.config.BridgeConstants;
@@ -25,12 +24,6 @@ import co.rsk.peg.utils.*;
 import co.rsk.peg.whitelist.LockWhitelist;
 import co.rsk.peg.whitelist.OneOffWhiteListEntry;
 import co.rsk.test.builders.BridgeSupportBuilder;
-import com.google.common.collect.Lists;
-import org.bouncycastle.asn1.x9.X9ECParameters;
-import org.bouncycastle.crypto.ec.CustomNamedCurves;
-import org.bouncycastle.crypto.params.ECDomainParameters;
-import org.bouncycastle.crypto.params.ECPrivateKeyParameters;
-import org.bouncycastle.crypto.signers.ECDSASigner;
 import org.bouncycastle.util.encoders.Hex;
 import org.ethereum.config.Constants;
 import org.ethereum.config.blockchain.upgrades.ActivationConfig;
@@ -42,9 +35,6 @@ import org.ethereum.core.Transaction;
 import org.ethereum.crypto.ECKey;
 import org.ethereum.crypto.HashUtil;
 import org.ethereum.util.ByteUtil;
-import org.ethereum.util.RLP;
-import org.ethereum.util.RLPList;
-import org.ethereum.vm.LogInfo;
 import org.ethereum.vm.PrecompiledContracts;
 import org.ethereum.vm.exception.VMException;
 import org.hamcrest.MatcherAssert;
@@ -63,10 +53,7 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static co.rsk.peg.PegTestUtils.*;
-import static org.hamcrest.Matchers.hasItem;
 import static org.hamcrest.Matchers.lessThanOrEqualTo;
-import static org.hamcrest.collection.IsCollectionWithSize.hasSize;
-import static org.hamcrest.collection.IsEmptyCollection.empty;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsNot.not;
 import static org.junit.jupiter.api.Assertions.*;
@@ -5303,607 +5290,6 @@ class BridgeSupportTest extends BridgeSupportTestBase {
     }
 
     @Test
-    void addSignature_fedPubKey_belongs_to_active_federation() throws Exception {
-        //Setup
-        FederationSupport mockFederationSupport = mock(FederationSupport.class);
-
-        // Creates new federation
-        List<BtcECKey> federation1Keys = Arrays.asList(
-            BtcECKey.fromPrivate(Hex.decode("fa01")),
-            BtcECKey.fromPrivate(Hex.decode("fa02"))
-        );
-        federation1Keys.sort(BtcECKey.PUBKEY_COMPARATOR);
-
-        Federation activeFederation = new Federation(
-            FederationTestUtils.getFederationMembersWithBtcKeys(federation1Keys),
-            Instant.ofEpochMilli(1000L),
-            0L,
-            btcRegTestParams
-        );
-
-        BridgeStorageProvider provider = mock(BridgeStorageProvider.class);
-        BridgeSupport bridgeSupport = new BridgeSupport(
-            bridgeConstantsRegtest,
-            provider,
-            mock(BridgeEventLogger.class),
-            new BtcLockSenderProvider(),
-            new PeginInstructionsProvider(),
-            mock(Repository.class),
-            mock(Block.class),
-            new Context(bridgeConstantsRegtest.getBtcParams()),
-            mockFederationSupport,
-            null,
-            null
-        );
-
-        when(mockFederationSupport.getActiveFederation()).thenReturn(activeFederation);
-        when(provider.getRskTxsWaitingForSignatures()).thenReturn(new TreeMap<>());
-
-        bridgeSupport.addSignature(BtcECKey.fromPrivate(Hex.decode("fa01")), null,
-            PegTestUtils.createHash3(1).getBytes());
-
-        verify(provider, times(1)).getRskTxsWaitingForSignatures();
-    }
-
-    @Test
-    void addSignature_fedPubKey_belongs_to_retiring_federation() throws Exception {
-        //Setup
-        FederationSupport mockFederationSupport = mock(FederationSupport.class);
-        BridgeStorageProvider provider = mock(BridgeStorageProvider.class);
-        BridgeSupport bridgeSupport = new BridgeSupport(
-            bridgeConstantsRegtest,
-            provider,
-            mock(BridgeEventLogger.class),
-            new BtcLockSenderProvider(),
-            new PeginInstructionsProvider(),
-            mock(Repository.class),
-            mock(Block.class),
-            new Context(bridgeConstantsRegtest.getBtcParams()),
-            mockFederationSupport,
-            null,
-            null
-        );
-
-        // Creates retiring federation
-        List<BtcECKey> federation1Keys = Arrays.asList(
-            BtcECKey.fromPrivate(Hex.decode("fa01")),
-            BtcECKey.fromPrivate(Hex.decode("fa02")));
-        federation1Keys.sort(BtcECKey.PUBKEY_COMPARATOR);
-
-        Federation retiringFederation = new Federation(
-            FederationTestUtils.getFederationMembersWithBtcKeys(federation1Keys),
-            Instant.ofEpochMilli(1000L),
-            0L,
-            btcRegTestParams
-        );
-
-        // Creates active federation
-        List<BtcECKey> activeFederationKeys = Arrays.asList(
-            BtcECKey.fromPrivate(Hex.decode("fa03")),
-            BtcECKey.fromPrivate(Hex.decode("fa04"))
-        );
-        activeFederationKeys.sort(BtcECKey.PUBKEY_COMPARATOR);
-
-        Federation activeFederation = new Federation(
-            FederationTestUtils.getFederationMembersWithBtcKeys(activeFederationKeys),
-            Instant.ofEpochMilli(1000L),
-            0L,
-            btcRegTestParams
-        );
-
-        when(mockFederationSupport.getActiveFederation()).thenReturn(activeFederation);
-        when(mockFederationSupport.getRetiringFederation()).thenReturn(retiringFederation);
-        when(provider.getRskTxsWaitingForSignatures()).thenReturn(new TreeMap<>());
-
-        bridgeSupport.addSignature(BtcECKey.fromPrivate(Hex.decode("fa01")), null,
-            PegTestUtils.createHash3(1).getBytes());
-
-        verify(provider, times(1)).getRskTxsWaitingForSignatures();
-    }
-
-    @Test
-    void addSignature_fedPubKey_no_belong_to_retiring_or_active_federation() throws Exception {
-        //Setup
-        FederationSupport mockFederationSupport = mock(FederationSupport.class);
-        BridgeStorageProvider provider = mock(BridgeStorageProvider.class);
-        BridgeSupport bridgeSupport = new BridgeSupport(
-            bridgeConstantsRegtest,
-            provider,
-            mock(BridgeEventLogger.class),
-            new BtcLockSenderProvider(),
-            new PeginInstructionsProvider(),
-            mock(Repository.class),
-            mock(Block.class),
-            new Context(bridgeConstantsRegtest.getBtcParams()),
-            mockFederationSupport,
-            null,
-            null
-        );
-
-        // Creates retiring federation
-        List<BtcECKey> federation1Keys = Arrays.asList(
-            BtcECKey.fromPrivate(Hex.decode("fa01")),
-            BtcECKey.fromPrivate(Hex.decode("fa02"))
-        );
-        federation1Keys.sort(BtcECKey.PUBKEY_COMPARATOR);
-
-        Federation retiringFederation = new Federation(
-            FederationTestUtils.getFederationMembersWithBtcKeys(federation1Keys),
-            Instant.ofEpochMilli(1000L),
-            0L,
-            btcRegTestParams
-        );
-
-        // Creates active federation
-        List<BtcECKey> activeFederationKeys = Arrays.asList(
-            BtcECKey.fromPrivate(Hex.decode("fa03")),
-            BtcECKey.fromPrivate(Hex.decode("fa04"))
-        );
-        activeFederationKeys.sort(BtcECKey.PUBKEY_COMPARATOR);
-
-        Federation activeFederation = new Federation(
-            FederationTestUtils.getFederationMembersWithBtcKeys(activeFederationKeys),
-            Instant.ofEpochMilli(1000L),
-            0L,
-            btcRegTestParams);
-
-        when(mockFederationSupport.getActiveFederation()).thenReturn(activeFederation);
-        when(mockFederationSupport.getRetiringFederation()).thenReturn(retiringFederation);
-
-        bridgeSupport.addSignature(BtcECKey.fromPrivate(Hex.decode("fa05")), null,
-            PegTestUtils.createHash3(1).getBytes());
-
-        verify(provider, times(0)).getRskTxsWaitingForSignatures();
-    }
-
-    @Test
-    void addSignature_fedPubKey_no_belong_to_active_federation_no_existing_retiring_fed() throws Exception {
-        BridgeStorageProvider provider = mock(BridgeStorageProvider.class);
-        BridgeSupport bridgeSupport = getBridgeSupport(
-            bridgeConstantsRegtest,
-            provider,
-            mock(Repository.class),
-            mock(BridgeEventLogger.class),
-            mock(Block.class),
-            null
-        );
-
-        bridgeSupport.addSignature(BtcECKey.fromPrivate(Hex.decode("fa03")), null,
-            PegTestUtils.createHash3(1).getBytes());
-
-        verify(provider, times(0)).getRskTxsWaitingForSignatures();
-    }
-
-    @Test
-    void addSignatureToMissingTransaction() throws Exception {
-        // Federation is the genesis federation ATM
-        Federation federation = bridgeConstantsRegtest.getGenesisFederation();
-        Repository repository = createRepository();
-
-        BridgeStorageProvider providerForSupport = new BridgeStorageProvider(
-            repository,
-            PrecompiledContracts.BRIDGE_ADDR,
-            bridgeConstantsRegtest,
-            activationsBeforeForks
-        );
-        BridgeSupport bridgeSupport = getBridgeSupport(bridgeConstantsRegtest, providerForSupport, repository,
-            mock(BridgeEventLogger.class), mock(Block.class), null);
-
-        bridgeSupport.addSignature(federation.getBtcPublicKeys().get(0), null, PegTestUtils.createHash().getBytes());
-        bridgeSupport.save();
-
-        BridgeStorageProvider provider = new BridgeStorageProvider(repository, PrecompiledContracts.BRIDGE_ADDR,
-            bridgeConstantsRegtest, activationsBeforeForks);
-
-        Assertions.assertTrue(provider.getRskTxsWaitingForSignatures().isEmpty());
-    }
-
-    @Test
-    void addSignatureFromInvalidFederator() throws Exception {
-        Repository repository = createRepository();
-
-        BridgeSupport bridgeSupport = getBridgeSupport(bridgeConstantsRegtest,
-            new BridgeStorageProvider(
-                repository,
-                PrecompiledContracts.BRIDGE_ADDR,
-                bridgeConstantsRegtest,
-                activationsBeforeForks),
-            repository,
-            mock(BridgeEventLogger.class),
-            mock(Block.class), null);
-
-        bridgeSupport.addSignature(new BtcECKey(), null, PegTestUtils.createHash().getBytes());
-        bridgeSupport.save();
-
-        BridgeStorageProvider provider = new BridgeStorageProvider(repository, PrecompiledContracts.BRIDGE_ADDR,
-            bridgeConstantsRegtest, activationsBeforeForks);
-
-        Assertions.assertTrue(provider.getRskTxsWaitingForSignatures().isEmpty());
-    }
-
-    @Test
-    void addSignatureWithInvalidSignature() throws Exception {
-        addSignatureFromValidFederator(Lists.newArrayList(new BtcECKey()), 1, true, false, "InvalidParameters");
-    }
-
-    @Test
-    void addSignatureWithLessSignaturesThanExpected() throws Exception {
-        List<BtcECKey> keys = Collections.singletonList(BridgeRegTestConstants.REGTEST_FEDERATION_PRIVATE_KEYS.get(0));
-        addSignatureFromValidFederator(keys, 0, true, false, "InvalidParameters");
-    }
-
-    @Test
-    void addSignatureWithMoreSignaturesThanExpected() throws Exception {
-        List<BtcECKey> keys = Collections.singletonList(BridgeRegTestConstants.REGTEST_FEDERATION_PRIVATE_KEYS.get(0));
-        addSignatureFromValidFederator(keys, 2, true, false, "InvalidParameters");
-    }
-
-    @Test
-    void addSignatureNonCanonicalSignature() throws Exception {
-        List<BtcECKey> keys = Collections.singletonList(BridgeRegTestConstants.REGTEST_FEDERATION_PRIVATE_KEYS.get(0));
-        addSignatureFromValidFederator(keys, 1, false, false, "InvalidParameters");
-    }
-
-    @Test
-    void addSignatureCreateEventLog_preRSKIP326() throws Exception {
-
-        // Setup
-        Federation federation = bridgeConstantsRegtest.getGenesisFederation();
-        Repository track = createRepository().startTracking();
-        BridgeStorageProvider provider = new BridgeStorageProvider(track, PrecompiledContracts.BRIDGE_ADDR, bridgeConstantsRegtest, activationsBeforeForks);
-
-        // Build prev btc tx
-        BtcTransaction prevTx = new BtcTransaction(btcRegTestParams);
-        TransactionOutput prevOut = new TransactionOutput(btcRegTestParams, prevTx, Coin.FIFTY_COINS, federation.getAddress());
-        prevTx.addOutput(prevOut);
-
-        // Build btc tx to be signed
-        BtcTransaction btcTx = new BtcTransaction(btcRegTestParams);
-        btcTx.addInput(prevOut).setScriptSig(createBaseInputScriptThatSpendsFromTheFederation(federation));
-        TransactionOutput output = new TransactionOutput(btcRegTestParams, btcTx, Coin.COIN, new BtcECKey().toAddress(btcRegTestParams));
-        btcTx.addOutput(output);
-
-        // Save btc tx to be signed
-        final Keccak256 rskTxHash = PegTestUtils.createHash3(1);
-        provider.getRskTxsWaitingForSignatures().put(rskTxHash, btcTx);
-        provider.save();
-        track.commit();
-
-        // Setup BridgeSupport
-        BridgeEventLogger eventLogger = mock(BridgeEventLogger.class);
-
-        ActivationConfig.ForBlock activations = mock(ActivationConfig.ForBlock.class);
-        when(activations.isActive(ConsensusRule.RSKIP326)).thenReturn(false);
-
-        BridgeSupport bridgeSupport = getBridgeSupport(
-            bridgeConstantsRegtest,
-                provider,
-                track,
-                eventLogger,
-                mock(Block.class),
-                null,
-                activations
-        );
-
-        // Create signed hash of Btc tx
-        Script inputScript = btcTx.getInputs().get(0).getScriptSig();
-        List<ScriptChunk> chunks = inputScript.getChunks();
-        byte[] program = chunks.get(chunks.size() - 1).data;
-        Script redeemScript = new Script(program);
-        Sha256Hash sigHash = btcTx.hashForSignature(0, redeemScript, BtcTransaction.SigHash.ALL, false);
-        BtcECKey privateKeyToSignWith = BridgeRegTestConstants.REGTEST_FEDERATION_PRIVATE_KEYS.get(0);
-
-        BtcECKey.ECDSASignature sig = privateKeyToSignWith.sign(sigHash);
-        List derEncodedSigs = Collections.singletonList(sig.encodeToDER());
-
-        BtcECKey federatorPubKey = findPublicKeySignedBy(federation.getBtcPublicKeys(), privateKeyToSignWith);
-        bridgeSupport.addSignature(federatorPubKey, derEncodedSigs, rskTxHash.getBytes());
-
-        verify(eventLogger, times(1)).logAddSignature(federatorPubKey, btcTx, rskTxHash.getBytes());
-    }
-
-    @Test
-    public void addSignature_invalidSignatureBeforeRSKIP326_eventStillEmitted() throws Exception {
-
-        // Setup
-        Federation federation = bridgeConstantsRegtest.getGenesisFederation();
-        Repository track = createRepository().startTracking();
-        BridgeStorageProvider provider = new BridgeStorageProvider(track, PrecompiledContracts.BRIDGE_ADDR, bridgeConstantsRegtest, activationsBeforeForks);
-
-        // Build prev btc tx
-        BtcTransaction prevTx = new BtcTransaction(btcRegTestParams);
-        TransactionOutput prevOut = new TransactionOutput(btcRegTestParams, prevTx, Coin.FIFTY_COINS, federation.getAddress());
-        prevTx.addOutput(prevOut);
-
-        // Build btc tx to be signed
-        BtcTransaction btcTx = new BtcTransaction(btcRegTestParams);
-        btcTx.addInput(prevOut).setScriptSig(createBaseInputScriptThatSpendsFromTheFederation(federation));
-        TransactionOutput output = new TransactionOutput(btcRegTestParams, btcTx, Coin.COIN, new BtcECKey().toAddress(btcRegTestParams));
-        btcTx.addOutput(output);
-
-        // Save btc tx to be signed
-        final Keccak256 rskTxHash = PegTestUtils.createHash3(1);
-        provider.getRskTxsWaitingForSignatures().put(rskTxHash, btcTx);
-        provider.save();
-        track.commit();
-
-        // Setup BridgeSupport
-        BridgeEventLogger eventLogger = mock(BridgeEventLogger.class);
-
-        ActivationConfig.ForBlock activations = mock(ActivationConfig.ForBlock.class);
-        when(activations.isActive(ConsensusRule.RSKIP326)).thenReturn(false);
-
-        BridgeSupport bridgeSupport = getBridgeSupport(
-                bridgeConstantsRegtest,
-                provider,
-                track,
-                eventLogger,
-                mock(Block.class),
-                null,
-                activations
-        );
-
-        BtcECKey privateKeyToSignWith = BridgeRegTestConstants.REGTEST_FEDERATION_PRIVATE_KEYS.get(0);
-
-        byte[] malformedSignature = new byte[70];
-        for (int i = 0; i < malformedSignature.length; i++) {
-            malformedSignature[i] = (byte) i;
-        }
-
-        List derEncodedSigs = Collections.singletonList(malformedSignature);
-
-        BtcECKey federatorPubKey = findPublicKeySignedBy(federation.getBtcPublicKeys(), privateKeyToSignWith);
-        bridgeSupport.addSignature(federatorPubKey, derEncodedSigs, rskTxHash.getBytes());
-
-        verify(eventLogger, times(1)).logAddSignature(federatorPubKey, btcTx, rskTxHash.getBytes());
-    }
-
-    @Test
-    public void addSignature_afterRSKIP326_eventEmitted() throws Exception {
-
-        // Setup
-        Federation federation = bridgeConstantsRegtest.getGenesisFederation();
-        Repository track = createRepository().startTracking();
-        BridgeStorageProvider provider = new BridgeStorageProvider(track, PrecompiledContracts.BRIDGE_ADDR, bridgeConstantsRegtest, activationsBeforeForks);
-
-        // Build prev btc tx
-        BtcTransaction prevTx = new BtcTransaction(btcRegTestParams);
-        TransactionOutput prevOut = new TransactionOutput(btcRegTestParams, prevTx, Coin.FIFTY_COINS, federation.getAddress());
-        prevTx.addOutput(prevOut);
-
-        // Build btc tx to be signed
-        BtcTransaction btcTx = new BtcTransaction(btcRegTestParams);
-        btcTx.addInput(prevOut).setScriptSig(createBaseInputScriptThatSpendsFromTheFederation(federation));
-        TransactionOutput output = new TransactionOutput(btcRegTestParams, btcTx, Coin.COIN, new BtcECKey().toAddress(btcRegTestParams));
-        btcTx.addOutput(output);
-
-        // Save btc tx to be signed
-        final Keccak256 rskTxHash = PegTestUtils.createHash3(1);
-        provider.getRskTxsWaitingForSignatures().put(rskTxHash, btcTx);
-        provider.save();
-        track.commit();
-
-        // Setup BridgeSupport
-        BridgeEventLogger eventLogger = mock(BridgeEventLogger.class);
-
-        ActivationConfig.ForBlock activations = mock(ActivationConfig.ForBlock.class);
-        when(activations.isActive(ConsensusRule.RSKIP326)).thenReturn(true);
-
-        BridgeSupport bridgeSupport = getBridgeSupport(
-                bridgeConstantsRegtest,
-                provider,
-                track,
-                eventLogger,
-                mock(Block.class),
-                null,
-                activations
-        );
-
-        // Create signed hash of Btc tx
-        Script inputScript = btcTx.getInputs().get(0).getScriptSig();
-        List<ScriptChunk> chunks = inputScript.getChunks();
-        byte[] program = chunks.get(chunks.size() - 1).data;
-        Script redeemScript = new Script(program);
-        Sha256Hash sigHash = btcTx.hashForSignature(0, redeemScript, BtcTransaction.SigHash.ALL, false);
-        BtcECKey privateKeyToSignWith = BridgeRegTestConstants.REGTEST_FEDERATION_PRIVATE_KEYS.get(0);
-
-        BtcECKey.ECDSASignature sig = privateKeyToSignWith.sign(sigHash);
-        List derEncodedSigs = Collections.singletonList(sig.encodeToDER());
-
-        BtcECKey federatorPubKey = findPublicKeySignedBy(federation.getBtcPublicKeys(), privateKeyToSignWith);
-        bridgeSupport.addSignature(federatorPubKey, derEncodedSigs, rskTxHash.getBytes());
-
-        verify(eventLogger, times(1)).logAddSignature(federatorPubKey, btcTx, rskTxHash.getBytes());
-    }
-
-    @Test
-    public void addSignature_invalidSignatureAfterRSKIP326_noEventEmitted() throws Exception {
-
-        // Setup
-        Federation federation = bridgeConstantsRegtest.getGenesisFederation();
-        Repository track = createRepository().startTracking();
-        BridgeStorageProvider provider = new BridgeStorageProvider(track, PrecompiledContracts.BRIDGE_ADDR, bridgeConstantsRegtest, activationsBeforeForks);
-
-        // Build prev btc tx
-        BtcTransaction prevTx = new BtcTransaction(btcRegTestParams);
-        TransactionOutput prevOut = new TransactionOutput(btcRegTestParams, prevTx, Coin.FIFTY_COINS, federation.getAddress());
-        prevTx.addOutput(prevOut);
-
-        // Build btc tx to be signed
-        BtcTransaction btcTx = new BtcTransaction(btcRegTestParams);
-        btcTx.addInput(prevOut).setScriptSig(createBaseInputScriptThatSpendsFromTheFederation(federation));
-        TransactionOutput output = new TransactionOutput(btcRegTestParams, btcTx, Coin.COIN, new BtcECKey().toAddress(btcRegTestParams));
-        btcTx.addOutput(output);
-
-        // Save btc tx to be signed
-        final Keccak256 rskTxHash = PegTestUtils.createHash3(1);
-        provider.getRskTxsWaitingForSignatures().put(rskTxHash, btcTx);
-        provider.save();
-        track.commit();
-
-        // Setup BridgeSupport
-        BridgeEventLogger eventLogger = mock(BridgeEventLogger.class);
-
-        ActivationConfig.ForBlock activations = mock(ActivationConfig.ForBlock.class);
-        when(activations.isActive(ConsensusRule.RSKIP326)).thenReturn(true);
-
-        BridgeSupport bridgeSupport = getBridgeSupport(
-                bridgeConstantsRegtest,
-                provider,
-                track,
-                eventLogger,
-                mock(Block.class),
-                null,
-                activations
-        );
-
-        BtcECKey privateKeyToSignWith = BridgeRegTestConstants.REGTEST_FEDERATION_PRIVATE_KEYS.get(0);
-
-        byte[] malformedSignature = new byte[70];
-        for (int i = 0; i < malformedSignature.length; i++) {
-            malformedSignature[i] = (byte) i;
-        }
-
-        List derEncodedSigs = Collections.singletonList(malformedSignature);
-
-        BtcECKey federatorPubKey = findPublicKeySignedBy(federation.getBtcPublicKeys(), privateKeyToSignWith);
-        bridgeSupport.addSignature(federatorPubKey, derEncodedSigs, rskTxHash.getBytes());
-
-        verify(eventLogger, times(0)).logAddSignature(federatorPubKey, btcTx, rskTxHash.getBytes());
-    }
-
-    @Test
-    void addSignatureTwice() throws Exception {
-        List<BtcECKey> keys = Collections.singletonList(BridgeRegTestConstants.REGTEST_FEDERATION_PRIVATE_KEYS.get(0));
-        addSignatureFromValidFederator(keys, 1, true, true, "PartiallySigned");
-    }
-
-    @Test
-    void addSignatureOneSignature() throws Exception {
-        List<BtcECKey> keys = Collections.singletonList(BridgeRegTestConstants.REGTEST_FEDERATION_PRIVATE_KEYS.get(0));
-        addSignatureFromValidFederator(keys, 1, true, false, "PartiallySigned");
-    }
-
-    @Test
-    void addSignatureTwoSignatures() throws Exception {
-        List<BtcECKey> federatorPrivateKeys = BridgeRegTestConstants.REGTEST_FEDERATION_PRIVATE_KEYS;
-        List<BtcECKey> keys = Arrays.asList(federatorPrivateKeys.get(0), federatorPrivateKeys.get(1));
-        addSignatureFromValidFederator(keys, 1, true, false, "FullySigned");
-    }
-
-    @Test
-    void addSignatureMultipleInputsPartiallyValid() throws Exception {
-        // Federation is the genesis federation ATM
-        Federation federation = bridgeConstantsRegtest.getGenesisFederation();
-        Repository repository = createRepository();
-
-        final Keccak256 keccak256 = PegTestUtils.createHash3(1);
-
-        BridgeStorageProvider provider = new BridgeStorageProvider(repository, PrecompiledContracts.BRIDGE_ADDR,
-            bridgeConstantsRegtest, activationsBeforeForks);
-
-        BtcTransaction prevTx1 = new BtcTransaction(btcRegTestParams);
-        TransactionOutput prevOut1 = new TransactionOutput(btcRegTestParams, prevTx1, Coin.FIFTY_COINS, federation.getAddress());
-        prevTx1.addOutput(prevOut1);
-        BtcTransaction prevTx2 = new BtcTransaction(btcRegTestParams);
-        TransactionOutput prevOut2 = new TransactionOutput(btcRegTestParams, prevTx1, Coin.FIFTY_COINS, federation.getAddress());
-        prevTx2.addOutput(prevOut2);
-        BtcTransaction prevTx3 = new BtcTransaction(btcRegTestParams);
-        TransactionOutput prevOut3 = new TransactionOutput(btcRegTestParams, prevTx1, Coin.FIFTY_COINS, federation.getAddress());
-        prevTx3.addOutput(prevOut3);
-
-        BtcTransaction t = new BtcTransaction(btcRegTestParams);
-        TransactionOutput output = new TransactionOutput(btcRegTestParams, t, Coin.COIN, new BtcECKey().toAddress(btcRegTestParams));
-        t.addOutput(output);
-        t.addInput(prevOut1).setScriptSig(createBaseInputScriptThatSpendsFromTheFederation(federation));
-        t.addInput(prevOut2).setScriptSig(createBaseInputScriptThatSpendsFromTheFederation(federation));
-        t.addInput(prevOut3).setScriptSig(createBaseInputScriptThatSpendsFromTheFederation(federation));
-        provider.getRskTxsWaitingForSignatures().put(keccak256, t);
-        provider.save();
-
-        ActivationConfig.ForBlock activations = mock(ActivationConfig.ForBlock.class);
-        List<LogInfo> logs = new ArrayList<>();
-        BridgeEventLogger eventLogger = new BrigeEventLoggerLegacyImpl(bridgeConstantsRegtest, activations, logs);
-        BridgeSupport bridgeSupport = getBridgeSupport(
-            bridgeConstantsRegtest,
-            new BridgeStorageProvider(
-                repository,
-                contractAddress,
-                bridgeConstantsRegtest,
-                activationsAfterForks
-            ),
-            repository,
-            eventLogger,
-            mock(Block.class),
-            null
-        );
-
-        // Generate valid signatures for inputs
-        List<byte[]> derEncodedSigsFirstFed = new ArrayList<>();
-        List<byte[]> derEncodedSigsSecondFed = new ArrayList<>();
-        BtcECKey privateKeyOfFirstFed = BridgeRegTestConstants.REGTEST_FEDERATION_PRIVATE_KEYS.get(0);
-        BtcECKey privateKeyOfSecondFed = BridgeRegTestConstants.REGTEST_FEDERATION_PRIVATE_KEYS.get(1);
-
-        BtcECKey.ECDSASignature lastSig = null;
-        for (int i = 0; i < 3; i++) {
-            Script inputScript = t.getInput(i).getScriptSig();
-            List<ScriptChunk> chunks = inputScript.getChunks();
-            byte[] program = chunks.get(chunks.size() - 1).data;
-            Script redeemScript = new Script(program);
-            Sha256Hash sighash = t.hashForSignature(i, redeemScript, BtcTransaction.SigHash.ALL, false);
-
-            // Sign the last input with a random key
-            // but keep the good signature for a subsequent call
-            BtcECKey.ECDSASignature sig = privateKeyOfFirstFed.sign(sighash);
-            if (i == 2) {
-                lastSig = sig;
-                sig = new BtcECKey().sign(sighash);
-            }
-            derEncodedSigsFirstFed.add(sig.encodeToDER());
-            derEncodedSigsSecondFed.add(privateKeyOfSecondFed.sign(sighash).encodeToDER());
-        }
-
-        // Sign with two valid signatures and one invalid signature
-        bridgeSupport.addSignature(findPublicKeySignedBy(federation.getBtcPublicKeys(), privateKeyOfFirstFed), derEncodedSigsFirstFed, keccak256.getBytes());
-        bridgeSupport.save();
-
-        // Sign with two valid signatures and one malformed signature
-        byte[] malformedSignature = new byte[lastSig.encodeToDER().length];
-        for (int i = 0; i < malformedSignature.length; i++) {
-            malformedSignature[i] = (byte) i;
-        }
-        derEncodedSigsFirstFed.set(2, malformedSignature);
-        bridgeSupport.addSignature(findPublicKeySignedBy(federation.getBtcPublicKeys(), privateKeyOfFirstFed), derEncodedSigsFirstFed, keccak256.getBytes());
-        bridgeSupport.save();
-
-        // Sign with fully valid signatures for same federator
-        derEncodedSigsFirstFed.set(2, lastSig.encodeToDER());
-        bridgeSupport.addSignature(findPublicKeySignedBy(federation.getBtcPublicKeys(), privateKeyOfFirstFed), derEncodedSigsFirstFed, keccak256.getBytes());
-        bridgeSupport.save();
-
-        // Sign with second federation
-        bridgeSupport.addSignature(findPublicKeySignedBy(federation.getBtcPublicKeys(), privateKeyOfSecondFed), derEncodedSigsSecondFed, keccak256.getBytes());
-        bridgeSupport.save();
-
-        provider = new BridgeStorageProvider(repository, PrecompiledContracts.BRIDGE_ADDR, bridgeConstantsRegtest, activationsBeforeForks);
-
-        Assertions.assertTrue(provider.getRskTxsWaitingForSignatures().isEmpty());
-        MatcherAssert.assertThat(logs, is(not(empty())));
-        MatcherAssert.assertThat(logs, hasSize(5));
-        LogInfo releaseTxEvent = logs.get(4);
-        MatcherAssert.assertThat(releaseTxEvent.getTopics(), hasSize(1));
-        MatcherAssert.assertThat(releaseTxEvent.getTopics(), hasItem(Bridge.RELEASE_BTC_TOPIC));
-        BtcTransaction releaseTx = new BtcTransaction(btcRegTestParams, ((RLPList) RLP.decode2(releaseTxEvent.getData()).get(0)).get(1).getRLPData());
-        // Verify all inputs fully signed
-        for (int i = 0; i < releaseTx.getInputs().size(); i++) {
-            Script retrievedScriptSig = releaseTx.getInput(i).getScriptSig();
-            Assertions.assertEquals(4, retrievedScriptSig.getChunks().size());
-            assertTrue(Objects.requireNonNull(retrievedScriptSig.getChunks().get(1).data).length > 0);
-            assertTrue(Objects.requireNonNull(retrievedScriptSig.getChunks().get(2).data).length > 0);
-        }
-    }
-
-    @Test
     void getTransactionType_pegin_tx() {
         BridgeSupport bridgeSupport = bridgeSupportBuilder
             .withBridgeConstants(bridgeConstantsRegtest)
@@ -7433,127 +6819,6 @@ class BridgeSupportTest extends BridgeSupportTestBase {
                 Assertions.assertEquals(i, input.getOutpoint().getIndex());
             }
         }
-    }
-
-    /**
-     * Helper method to test addSignature() with a valid federatorPublicKey parameter and both valid/invalid signatures
-     *
-     * @param privateKeysToSignWith keys used to sign the tx. Federator key when we want to produce a valid signature, a random key when we want to produce an invalid signature
-     * @param numberOfInputsToSign  There is just 1 input. 1 when testing the happy case, other values to test attacks/bugs.
-     * @param signatureCanonical    Signature should be canonical. true when testing the happy case, false to test attacks/bugs.
-     * @param signTwice             Sign again with the same key
-     * @param expectedResult        "InvalidParameters", "PartiallySigned" or "FullySigned"
-     */
-    private void addSignatureFromValidFederator(List<BtcECKey> privateKeysToSignWith, int numberOfInputsToSign, boolean signatureCanonical, boolean signTwice, String expectedResult) throws Exception {
-        // Federation is the genesis federation ATM
-        Federation federation = bridgeConstantsRegtest.getGenesisFederation();
-        Repository repository = createRepository();
-
-        final Keccak256 keccak256 = PegTestUtils.createHash3();
-
-        Repository track = repository.startTracking();
-        BridgeStorageProvider provider = new BridgeStorageProvider(track, PrecompiledContracts.BRIDGE_ADDR, bridgeConstantsRegtest, activationsBeforeForks);
-
-        BtcTransaction prevTx = new BtcTransaction(btcRegTestParams);
-        TransactionOutput prevOut = new TransactionOutput(btcRegTestParams, prevTx, Coin.FIFTY_COINS, federation.getAddress());
-        prevTx.addOutput(prevOut);
-
-        BtcTransaction t = new BtcTransaction(btcRegTestParams);
-        TransactionOutput output = new TransactionOutput(btcRegTestParams, t, Coin.COIN, new BtcECKey().toAddress(btcRegTestParams));
-        t.addOutput(output);
-        t.addInput(prevOut).setScriptSig(createBaseInputScriptThatSpendsFromTheFederation(federation));
-        provider.getRskTxsWaitingForSignatures().put(keccak256, t);
-        provider.save();
-        track.commit();
-
-        track = repository.startTracking();
-        ActivationConfig.ForBlock activations = mock(ActivationConfig.ForBlock.class);
-        List<LogInfo> logs = new ArrayList<>();
-        BridgeEventLogger eventLogger = new BrigeEventLoggerLegacyImpl(bridgeConstantsRegtest, activations, logs);
-        BridgeSupport bridgeSupport = getBridgeSupport(
-            bridgeConstantsRegtest,
-            new BridgeStorageProvider(
-                track,
-                contractAddress,
-                bridgeConstantsRegtest,
-                activationsAfterForks
-            ),
-            track,
-            eventLogger,
-            mock(Block.class),
-            null
-        );
-
-        Script inputScript = t.getInputs().get(0).getScriptSig();
-        List<ScriptChunk> chunks = inputScript.getChunks();
-        byte[] program = chunks.get(chunks.size() - 1).data;
-        Script redeemScript = new Script(program);
-        Sha256Hash sighash = t.hashForSignature(0, redeemScript, BtcTransaction.SigHash.ALL, false);
-
-        BtcECKey.ECDSASignature sig = privateKeysToSignWith.get(0).sign(sighash);
-        if (!signatureCanonical) {
-            sig = new BtcECKey.ECDSASignature(sig.r, BtcECKey.CURVE.getN().subtract(sig.s));
-        }
-        byte[] derEncodedSig = sig.encodeToDER();
-
-        List derEncodedSigs = new ArrayList();
-        for (int i = 0; i < numberOfInputsToSign; i++) {
-            derEncodedSigs.add(derEncodedSig);
-        }
-        bridgeSupport.addSignature(findPublicKeySignedBy(federation.getBtcPublicKeys(), privateKeysToSignWith.get(0)), derEncodedSigs, keccak256.getBytes());
-        if (signTwice) {
-            // Create another valid signature with the same private key
-            ECDSASigner signer = new ECDSASigner();
-            X9ECParameters CURVE_PARAMS = CustomNamedCurves.getByName("secp256k1");
-            ECDomainParameters CURVE = new ECDomainParameters(CURVE_PARAMS.getCurve(), CURVE_PARAMS.getG(), CURVE_PARAMS.getN(), CURVE_PARAMS.getH());
-            ECPrivateKeyParameters privKey = new ECPrivateKeyParameters(privateKeysToSignWith.get(0).getPrivKey(), CURVE);
-            signer.init(true, privKey);
-            BigInteger[] components = signer.generateSignature(sighash.getBytes());
-            BtcECKey.ECDSASignature sig2 = new BtcECKey.ECDSASignature(components[0], components[1]).toCanonicalised();
-            bridgeSupport.addSignature(findPublicKeySignedBy(federation.getBtcPublicKeys(), privateKeysToSignWith.get(0)), Lists.newArrayList(sig2.encodeToDER()), keccak256.getBytes());
-        }
-        if (privateKeysToSignWith.size() > 1) {
-            BtcECKey.ECDSASignature sig2 = privateKeysToSignWith.get(1).sign(sighash);
-            byte[] derEncodedSig2 = sig2.encodeToDER();
-            List derEncodedSigs2 = new ArrayList();
-            for (int i = 0; i < numberOfInputsToSign; i++) {
-                derEncodedSigs2.add(derEncodedSig2);
-            }
-            bridgeSupport.addSignature(findPublicKeySignedBy(federation.getBtcPublicKeys(), privateKeysToSignWith.get(1)), derEncodedSigs2, keccak256.getBytes());
-        }
-        bridgeSupport.save();
-        track.commit();
-
-        provider = new BridgeStorageProvider(repository, PrecompiledContracts.BRIDGE_ADDR, bridgeConstantsRegtest, activationsBeforeForks);
-
-        if ("FullySigned".equals(expectedResult)) {
-            Assertions.assertTrue(provider.getRskTxsWaitingForSignatures().isEmpty());
-            MatcherAssert.assertThat(logs, is(not(empty())));
-            MatcherAssert.assertThat(logs, hasSize(3));
-            LogInfo releaseTxEvent = logs.get(2);
-            MatcherAssert.assertThat(releaseTxEvent.getTopics(), hasSize(1));
-            MatcherAssert.assertThat(releaseTxEvent.getTopics(), hasItem(Bridge.RELEASE_BTC_TOPIC));
-            BtcTransaction releaseTx = new BtcTransaction(btcRegTestParams, ((RLPList) RLP.decode2(releaseTxEvent.getData()).get(0)).get(1).getRLPData());
-            Script retrievedScriptSig = releaseTx.getInput(0).getScriptSig();
-            Assertions.assertEquals(4, retrievedScriptSig.getChunks().size());
-            assertTrue(retrievedScriptSig.getChunks().get(1).data.length > 0);
-            assertTrue(retrievedScriptSig.getChunks().get(2).data.length > 0);
-        } else {
-            Script retrievedScriptSig = provider.getRskTxsWaitingForSignatures().get(keccak256).getInput(0).getScriptSig();
-            Assertions.assertEquals(4, retrievedScriptSig.getChunks().size());
-            boolean expectSignatureToBePersisted = "PartiallySigned".equals(expectedResult); // for "InvalidParameters"
-            Assertions.assertEquals(expectSignatureToBePersisted, retrievedScriptSig.getChunks().get(1).data.length > 0);
-            assertFalse(retrievedScriptSig.getChunks().get(2).data.length > 0);
-        }
-    }
-
-    private BtcECKey findPublicKeySignedBy(List<BtcECKey> pubs, BtcECKey pk) {
-        for (BtcECKey pub : pubs) {
-            if (Arrays.equals(pk.getPubKey(), pub.getPubKey())) {
-                return pub;
-            }
-        }
-        return pk;
     }
 
     private BridgeSupport getBridgeSupport(BridgeConstants constants, BridgeStorageProvider provider) {
